Declutter code
